+++ css/projects-section.css
@@ -73,7 +73,7 @@
 }
 
 /* Carousel code */
-#carousel {
+.carousel {
   position: relative;
   width: 600px;
   height: 600px;
@@ -100,18 +100,18 @@
   object-fit: contain; /* Maintain aspect ratio */
 }
 
+
 .carousel-image.active {
   opacity: 1; /* Show active image */
 }
-
-#carousel-indicators {
+.carousel-indicators {
   position: absolute;
   bottom: 0;
   width: 100%;
   text-align: center;
 }
 
-#carousel-indicators .indicator {
+.carousel-indicators .indicator {
   display: inline-block;
   margin: 0 5px;
   width: 20px;
@@ -122,14 +122,15 @@
   cursor: pointer;
 }
 
-#carousel-indicators .indicator.active {
+.carousel-indicators .indicator.active {
   background-color: #ff5400;
   width: 60px;
+  height: 20px;
   border-radius: 10px;
   cursor: pointer;
 }
 
-#carousel-indicators .indicator:hover {
+.carousel-indicators .indicator:hover {
   scale: 1.3;
   margin: 0 1rem;
   background-color: darkorange;
