+++ index.html
@@ -109,15 +109,15 @@
           <p>PROJECTS</p>
           <div class="bar-horizontal-small"></div>
 
-          <div class="project-container">
-            <div id="carousel">
+          <div class="project-container" data-carousel="proserv-carousel">
+            <div class="carousel">
               <div class="carousel-images">
                 <img class="carousel-image" src="./public/github_logo.png" />
                 <img class="carousel-image" src="./public/vite.svg" />
                 <img class="carousel-image" src="./public/ProServ-Mobile.png" />
                 <!-- Add more images as needed -->
               </div>
-              <div id="carousel-indicators">
+              <div class="carousel-indicators">
                 <!-- Indicators will be dynamically added here with JavaScript -->
               </div>
             </div>
@@ -155,16 +155,16 @@
             </div>
           </div>
 
-          <div class="project-container">
-            <div id="carousel">
-              <div class="carousel-i mages">
+          <div class="project-container" data-carousel="superweather-carousel">
+            <div class="carousel">
+              <div class="carousel-images">
                 <img class="carousel-image" src="./public/github_logo.png" />
                 <img class="carousel-image" src="./public/vite.svg" />
                 <img class="carousel-image" src="./public/ProServ-Mobile.png" />
-                <!-- Add more images as needed -->
+                
               </div>
-              <div id="carousel-indicators">
-                <!-- Indicators will be dynamically added here with JavaScript -->
+              <div class="carousel-indicators">
+                <!-- ... -->
               </div>
             </div>
